requirements.txt
# Core
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-dateutil>=2.8.2
pytz>=2023.3
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic>=1.10.7
python-dotenv>=1.0.0
redis>=4.5.0
aioredis>=2.0.0
aiohttp>=3.8.4

# Streamlit et visualisation
streamlit>=1.32.0
plotly>=5.15.0
seaborn>=0.13.0
matplotlib>=3.8.0

# Utilitaires
python-dotenv>=1.0.0
joblib>=1.3.0
tabula-py>=2.10.0

# Tests
pytest>=7.4.0
pytest-cov>=4.1.0